+++ beetsplug/fuzzy_search.py
@@ -16,7 +16,7 @@
 """
 import beets
 from beets.plugins import BeetsPlugin
-from beets.ui import Subcommand, decargs, print_
+from beets.ui import Subcommand, decargs, print_obj
 from beets.util.functemplate import Template
 import difflib
 
@@ -24,19 +24,18 @@ import difflib
 # THRESHOLD = 0.7
 
 
-def fuzzy_score(query, item):
-    return difflib.SequenceMatcher(a=query, b=item).quick_ratio()
+def fuzzy_score(queryMatcher, item):
+    queryMatcher.set_seq1(item)
+    return queryMatcher.quick_ratio()
 
 
-def is_match(query, item, album=False, verbose=False, threshold=0.7):
-    query = ' '.join(query)
-
+def is_match(queryMatcher, item, album=False, verbose=False, threshold=0.7):
     if album:
         values = [item.albumartist, item.album]
     else:
         values = [item.artist, item.album, item.title]
 
-    s = max(fuzzy_score(query.lower(), i.lower()) for i in values)
+    s = max(fuzzy_score(queryMatcher, i.lower()) for i in values)
     if verbose:
         return (s >= threshold, s)
     else:
@@ -45,36 +44,32 @@ def is_match(query, item, album=False, verbose=False, threshold=0.7):
 
 def fuzzy_list(lib, config, opts, args):
     query = decargs(args)
-    fmt = opts.format
+    query = ' '.join(query).lower()
+    queryMatcher = difflib.SequenceMatcher(b=query)
+
     if opts.threshold is not None:
         threshold = float(opts.threshold)
     else:
         threshold = float(conf['threshold'])
 
-    if fmt is None:
-        # If no specific template is supplied, use a default
-        if opts.album:
-            fmt = u'$albumartist - $album'
-        else:
-            fmt = u'$artist - $album - $title'
-    template = Template(fmt)
+    if opts.path:
+        fmt = '$path'
+    else:
+        fmt = opts.format
+    template = Template(fmt) if fmt else None
 
     if opts.album:
         objs = lib.albums()
     else:
         objs = lib.items()
 
-    items = filter(lambda i: is_match(query, i, album=opts.album,
+    items = filter(lambda i: is_match(queryMatcher, i, album=opts.album,
                                       threshold=threshold), objs)
-    for i in items:
-        if opts.path:
-            print_(i.item_dir() if opts.album else i.path)
-        elif opts.album:
-            print_(i.evaluate_template(template))
-        else:
-            print_(i.evaluate_template(template, lib))
+
+    for item in items:
+        print_obj(item, lib, config, template)
         if opts.verbose:
-            print(is_match(query, i, album=opts.album, verbose=True)[1])
+            print(is_match(queryMatcher, i, album=opts.album, verbose=True)[1])
 
 
 fuzzy_cmd = Subcommand('fuzzy',